試し

+++ BSClient/Player.cpp
@@ -92,7 +92,7 @@ void Player::Update()
 			float angle = -(bullet->GetAngle());
 			float time = bullet->GetBulletTime();
 			long sendData[5] = { htonl(type), htonl(bPos.x),htonl(bPos.y),htonl(angle),htonl(time) };*/
-			long sendData[2] = { htonl(type), htonl(rAngle)};
+			float sendData[2] = { htonl(type), htonl(rAngle)};
 			//long sendData = htonl(type);
 			int ret = NetWorkSendUDP(sock_, ip_, 8888, &sendData, sizeof(sendData));
 		}
